move deleteBranch into a git helper wrapper and add listBranches
